Moved the configuration capability into the common library and added config capability to WebApi
Moved the configuration capability into the common library and added config capability to WebApi as part of #21
