+++ src/dotnet/CarbonAware.CLI/CarbonAwareCLI.cs
@@ -1,5 +1,5 @@
 ﻿using CarbonAware;
-using CarbonAwareCLI.Config;
+using CarbonAware.Config;
 using CarbonAwareCLI.Options;
 using CommandLine;
 using Microsoft.Extensions.Configuration;
@@ -15,14 +15,11 @@ namespace CarbonAwareCLI
 {
     public class CarbonAwareCLI
     {
-        private const string CONFIG_SECTION_SERVICE_REGISTRATIONS = "service-registrations";
-        private const string CONFIG_SERVICES_ARRAY = "services";
-        private const string PLUGINS_FOLDER = "plugins";
+        
 
         private CarbonAwareCLIState _state { get; set; } = new CarbonAwareCLIState();
         private CarbonAwareCore _carbonAwareCore;
-        private ServiceProvider _serviceProvider;
-        private CompositionContainer _container;
+        private ServiceManager _serviceManager;
 
         /// <summary>
         /// Indicates if the command line arguments have been parsed successfully 
@@ -40,9 +37,9 @@ namespace CarbonAwareCLI
                 parseResult.WithNotParsed(ThrowOnParseError);
 
                 // Configure the services
-                ConfigureServices();
+                _serviceManager = new ServiceManager(_state.ConfigPath);
 
-                var plugin = _serviceProvider.GetService<ICarbonAwarePlugin>();
+                var plugin = _serviceManager.ServiceProvider.GetService<ICarbonAwarePlugin>();
 
                 // Create the new core using the plugin
                 _carbonAwareCore = new CarbonAwareCore(plugin);
@@ -56,89 +53,7 @@ namespace CarbonAwareCLI
             }
         }
 
-        /// <summary>
-        /// Configures the service provider 
-        /// </summary>
-        private void ConfigureServices()
-        {
-            var builder = new ConfigurationBuilder()
-                    .SetBasePath(Directory.GetCurrentDirectory())
-                    .AddJsonFile(_state.ConfigPath);
-
-            var config = builder.Build();
-            var section = config.GetSection(CONFIG_SECTION_SERVICE_REGISTRATIONS);
-            var services = section.GetSection(CONFIG_SERVICES_ARRAY).Get<List<ServiceRegistration>>();
-                
-            ValidateServiceSyntax(services);
-
-            LoadPluginAssemblies();
-
-            var serviceCollection = new ServiceCollection()
-                .AddLogging();
-
-            // Register Services
-            foreach (var service in services)
-            {
-                AddService(serviceCollection, service);
-            }
-
-            _serviceProvider = serviceCollection.BuildServiceProvider();
-
-            // Configure Services
-            foreach (var service in services)
-            {
-                var serviceType = Type.GetType(service.service, true);
-                var registeredService = _serviceProvider.GetService(serviceType) as IConfigurable;
-                var registeredServiceName = registeredService.GetType().Name;
-                var configSection = config.GetSection(registeredServiceName);
-                registeredService.Configure(configSection);
-            }
-        }
-
-
-        private void LoadPluginAssemblies()
-        {
-            // An aggregate catalog that combines multiple catalogs.
-            var catalog = new AggregateCatalog();
-            // Adds all the parts found in the same assembly as the Program class.
-            catalog.Catalogs.Add(new AssemblyCatalog(typeof(Program).Assembly));
-            
-            // Add all the parts found in the "plugins" folder
-            catalog.Catalogs.Add(new DirectoryCatalog(AppDomain.CurrentDomain.BaseDirectory + PLUGINS_FOLDER));
-
-            // Create the CompositionContainer with the parts in the catalog.
-            _container = new CompositionContainer(catalog);
-            _container.ComposeParts(this);
-
-            // Plugins folder DLL's are now loaded!
-        }
-
-        private void ValidateServiceSyntax(List<ServiceRegistration> services)
-        {
-            if (services == null) throw new ArgumentException($"Configuration file '{_state.ConfigPath}' is invalid.  Could not find services.");
-            foreach (var service in services)
-            {
-                if (service.service is null || service.implementation is null)
-                {
-                    throw new ArgumentException($"Service configuration is invalid.  Service: '{service.service}', Implementation: '{service.implementation}'");
-                }
-            }
-        }
-
-        private static void AddService(IServiceCollection serviceCollection, ServiceRegistration service)
-        {
-            try
-            {
-                var serviceType = Type.GetType(service.service, true);
-                var serviceImplementation = Type.GetType(service.implementation);
-
-                serviceCollection.AddSingleton(serviceType, serviceImplementation);
-            }
-            catch
-            {
-                throw new ArgumentException($"Error configuring service '${service.service}' with type $'{service.implementation}'");
-            }
-        }
+        
 
         /// <summary>
         /// Handles missing messages.  Currently reports the message tag as an argument exception.
