+++ src/dotnet/CarbonAware.Plugins.BasicJsonPlugin/CarbonAware.Plugins.BasicJsonPlugin.csproj
@@ -20,4 +20,8 @@
     <ProjectReference Include="..\CarbonAware\CarbonAware.csproj" />
   </ItemGroup>
 
+  <Target Name="PostBuild" AfterTargets="PostBuildEvent">
+    <Exec Command="copy /Y &quot;$(TargetDir)$(TargetName).dll&quot; &quot;$(SolutionDir)CarbonAware.CLI\bin\Debug\net6.0\plugins\$(TargetName).dll&quot;&#xD;&#xA;copy /Y &quot;$(TargetDir)$(TargetName).dll&quot; &quot;$(SolutionDir)CarbonAware.WebApi\plugins\$(TargetName).dll&quot;&#xD;&#xA;" />
+  </Target>
+
 </Project>
