+++ src/dotnet/CarbonAware.WebApi/Properties/launchSettings.json
@@ -5,7 +5,7 @@
     "anonymousAuthentication": true,
     "iisExpress": {
       "applicationUrl": "http://localhost:16911",
-      "sslPort": 44324
+      "sslPort": 0
     }
   },
   "profiles": {
@@ -32,7 +32,7 @@
       "launchBrowser": true,
       "launchUrl": "{Scheme}://{ServiceHost}:{ServicePort}/swagger",
       "publishAllPorts": true,
-      "useSSL": true
+      "useSSL": false
     }
   }
 }
