+++ src/dotnet/CarbonAware.WebApi/carbon-aware.json
@@ -1,0 +1,20 @@
+{
+  "service-registrations": {
+    "services": [
+      {
+        "service": "CarbonAware.ICarbonAwareStaticDataService, CarbonAware, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
+        "implementation": "CarbonAware.Plugins.BasicJsonPlugin.CarbonAwareStaticJsonDataService, CarbonAware.Plugins.BasicJsonPlugin, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
+      },
+      {
+        "service": "CarbonAware.ICarbonAwarePlugin, CarbonAware, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
+        "implementation": "CarbonAware.Plugins.BasicJsonPlugin.CarbonAwareBasicDataPlugin, CarbonAware.Plugins.BasicJsonPlugin, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
+      }
+    ]
+  },
+  "CarbonAwareStaticJsonDataService": {
+    "data-file": "data-files/dummy-data-azure-emissions.json"
+  },
+  "CarbonAwareBasicDataPlugin": {
+    // nothing 
+  }
+}
