+++ src/dotnet/CarbonAware/Config/ServiceRegistration.cs
@@ -1,4 +1,4 @@
-﻿namespace CarbonAwareCLI.Config
+﻿namespace CarbonAware.Config
 {
     public class ServiceRegistration
     {
